feat(ui): rebuild settings as sidebar cockpit with mobile drill-in (5a/5b)
Replaces the wrapping tab strip + stacked panels with the design handoff's
5a/5b layout: a 780px summoned dialog with a 176px nav rail (search field,
glyph section items with amber active indicator, live footer) and an
aligned label/control content pane on desktop; a full-screen section list
with current-value summaries and drill-in detail pages on mobile.

- SettingsShell owns the chrome and navigation; the six section panels stay
  mounted and toggle via hidden (stable ids, steers drafts survive).
- Coding CLI and Session bodies are extracted out of the Settings monolith
  into SettingsCodingCliPanel / SettingsSessionPanel, which seed their
  section-scoped state from the parent's single getSettings() payload; the
  provider/model trio, fable, and reduced-push stay in Settings.svelte
  (cross-section consumers + tested save/revert contracts).
- SettingsDefaultEnvironment is folded into the Coding CLI panel's GLOBAL
  DEFAULTS rows (data-testid + aria contracts preserved).
- Search: / focuses, Esc clears; filters the rail with amber per-section
  match badges, shows the header count, highlights matches amber-24%, and
  auto-expands matching collapsed groups. Counts and highlights derive from
  one shared index (settings-search.ts).
- Diagnostics issues drive the rail's red dot / the mobile 1-ISSUE chip;
  the rail footer wires the real app version + WS liveness.
- Deep links (diagnostics dot, plugin focus, steer edit, workspace picker)
  drill straight into the section detail on mobile; a plain gear open shows
  the list. settingsTab can now target codingAgents.
- Settings apply immediately as before; a static hint says so.
