+++ ui/messages/de.json
@@ -2235,8 +2235,8 @@
   "settings_default_agent_provider_title": "Standard-Coding-CLI",
   "settings_default_agent_provider_hint": "Wählt die interaktive CLI, die bei neuen Aufgaben vorausgewählt ist. Claude Code behält die gesamte Shepherd-Automation; Codex startet als steuerbare Terminal-Session, während CLI-spezifische Automation noch verdrahtet wird.",
   "settings_default_agent_provider_save_failed": "Standard-Coding-CLI konnte nicht geändert werden. Erneut versuchen.",
-  "settings_upnext_skip_cli_picker_label": "Coding-CLI-Auswahl in „Als Nächstes“ überspringen",
-  "settings_upnext_skip_cli_picker_hint": "Wenn aktiviert, startet der Schnellstart in „Als Nächstes“ sofort mit deiner Standard-Coding-CLI, ohne nachzufragen, welche CLI verwendet werden soll — auch wenn mehr als eine installiert ist.",
+  "settings_upnext_skip_cli_picker_label": "CLI-Auswahl in Up Next überspringen",
+  "settings_upnext_skip_cli_picker_hint": "Schnellstart startet ohne Nachfrage mit der Standard-CLI — auch wenn mehrere installiert sind.",
   "settings_upnext_skip_cli_picker_save_failed": "Einstellung für die „Als Nächstes“-Auswahl konnte nicht geändert werden. Erneut versuchen.",
   "settings_cli_claude_title": "Claude Code",
   "settings_cli_claude_hint": "Claude-basierte Läufe verwenden diese Einstellungen für Aufwand, Sprache und Authentifizierung.",
@@ -3200,7 +3200,7 @@
   "feat_build_queue_progress_toggle_title": "Build-Queue direkt öffnen",
   "feat_build_queue_progress_toggle_body": "Klicke auf das Fortschritts-Badge einer Session, um sie auszuwählen und ihre Build-Queue über dem Terminal zu öffnen. Klicke erneut auf das Badge der ausgewählten Session, um die Queue ein- oder auszuklappen.",
   "settings_default_codex_model_title": "Codex-Modell",
-  "settings_default_codex_model_hint": "Legt das Codex-Modell für neue Aufgaben und Codex-basierte autonome Starts fest. Standard lässt die lokale Codex-CLI ohne --model-Flag wählen.",
+  "settings_default_codex_model_hint": "Default lässt die lokale CLI ohne --model-Flag. Gepinnte Modelle gelten auch für Drain- und Autopilot-Starts.",
   "settings_default_codex_model_save_failed": "Das Standard-Codex-Modell konnte nicht geändert werden. Erneut versuchen.",
   "settings_default_environment_title": "Standard-Coding-Umgebung",
   "settings_default_environment_hint": "Wähle CLI und Modell für neue Aufgaben und unbeaufsichtigte Arbeit. Shepherd merkt sich je ein Modell für Claude Code und Codex.",
@@ -3376,5 +3376,22 @@
   "feat_codex_release_notes_title": "Codex-Änderungen vor dem Update lesen",
   "feat_codex_release_notes_body": "Das Codex-Update-Fenster zeigt jetzt die originalen GitHub-Release-Notes für jede auf npm veröffentlichte stabile Version in deinem Update-Bereich – neueste zuerst.",
   "feat_repo_owner_picker_title": "Das richtige Repository wählen",
-  "feat_repo_owner_picker_body": "Der Repository-Auswähler zeigt jetzt Owner und Repository und lässt nach Ownern suchen. Gleichnamige Repositories in verschiedenen Accounts lassen sich so leicht unterscheiden."
+  "feat_repo_owner_picker_body": "Der Repository-Auswähler zeigt jetzt Owner und Repository und lässt nach Ownern suchen. Gleichnamige Repositories in verschiedenen Accounts lassen sich so leicht unterscheiden.",
+  "settings_default_cli_desc": "Für neue Tasks und unbeaufsichtigte Läufe.",
+  "settings_search_placeholder": "Suchen",
+  "settings_search_placeholder_mobile": "Einstellungen durchsuchen",
+  "settings_search_aria": "Einstellungen durchsuchen",
+  "settings_search_matches": "{count} Treffer für \"{query}\"",
+  "settings_search_match_one": "1 Treffer für \"{query}\"",
+  "settings_apply_hint": "Änderungen gelten sofort",
+  "settings_group_count": "{count} Einstellungen",
+  "settings_group_count_one": "1 Einstellung",
+  "settings_live": "live",
+  "settings_offline": "offline",
+  "settings_back_aria": "Zurück",
+  "settings_diag_issue_one": "1 Problem",
+  "settings_diag_issues": "{count} Probleme",
+  "settings_plugins_installed_summary": "{count} installiert",
+  "feat_settings_redesign_title": "Einstellungen, neu gestaltet",
+  "feat_settings_redesign_body": "Die Einstellungen öffnen sich jetzt als Cockpit: Sektionsleiste links (am Telefon eine Vollbild-Liste mit Detailseiten), ausgerichtete Einstellungszeilen und ein Suchfeld — mit / erreichbar —, das Sektionen filtert und Treffer hervorhebt. Änderungen gelten weiterhin sofort."
 }
