+++ ui/messages/en.json
@@ -2235,8 +2235,8 @@
   "settings_default_agent_provider_title": "Default coding CLI",
   "settings_default_agent_provider_hint": "Chooses the interactive CLI preselected for new tasks. Claude Code keeps all Shepherd automation; Codex starts as a steerable terminal session while CLI-specific automation is still being wired.",
   "settings_default_agent_provider_save_failed": "Couldn't change default coding CLI. Retry.",
-  "settings_upnext_skip_cli_picker_label": "Skip the coding-CLI picker in Up Next",
-  "settings_upnext_skip_cli_picker_hint": "When enabled, Up Next quick-start launches with your default coding CLI immediately, without asking which CLI to use — even when more than one is installed.",
+  "settings_upnext_skip_cli_picker_label": "Skip CLI picker in Up Next",
+  "settings_upnext_skip_cli_picker_hint": "Quick-start launches with the default CLI without asking, even when several are installed.",
   "settings_upnext_skip_cli_picker_save_failed": "Couldn't change the Up Next picker setting. Retry.",
   "settings_cli_claude_title": "Claude Code",
   "settings_cli_claude_hint": "Claude-backed runs use these effort, language, and authentication settings.",
@@ -3200,7 +3200,7 @@
   "feat_build_queue_progress_toggle_title": "Open the build queue from progress",
   "feat_build_queue_progress_toggle_body": "Click a session's progress badge to select it and open its build queue above the terminal. Click the selected session's badge again to collapse or expand the queue.",
   "settings_default_codex_model_title": "Codex model",
-  "settings_default_codex_model_hint": "Sets the Codex model for new tasks and Codex-backed autonomous starts. Default lets the local Codex CLI choose without a --model flag.",
+  "settings_default_codex_model_hint": "Default leaves the local CLI unflagged. Pinned models also apply to drain and autopilot starts.",
   "settings_default_codex_model_save_failed": "Couldn't change the default Codex model. Retry.",
   "settings_default_environment_title": "Default coding environment",
   "settings_default_environment_hint": "Choose the CLI and its model for new tasks and unattended work. Shepherd remembers a separate model for Claude Code and Codex.",
@@ -3376,5 +3376,22 @@
   "feat_codex_release_notes_title": "Read Codex changes before updating",
   "feat_codex_release_notes_body": "The Codex update dialog now shows the original GitHub release notes for every npm-published stable version in your update range, newest first.",
   "feat_repo_owner_picker_title": "Pick the right repository",
-  "feat_repo_owner_picker_body": "The repository picker now shows each owner and repository, and lets you search by owner. Repositories with the same name in different accounts are easy to tell apart."
+  "feat_repo_owner_picker_body": "The repository picker now shows each owner and repository, and lets you search by owner. Repositories with the same name in different accounts are easy to tell apart.",
+  "settings_default_cli_desc": "Used for new tasks and unattended runs.",
+  "settings_search_placeholder": "Search",
+  "settings_search_placeholder_mobile": "Search settings",
+  "settings_search_aria": "Search settings",
+  "settings_search_matches": "{count} matches for \"{query}\"",
+  "settings_search_match_one": "1 match for \"{query}\"",
+  "settings_apply_hint": "changes apply immediately",
+  "settings_group_count": "{count} settings",
+  "settings_group_count_one": "1 setting",
+  "settings_live": "live",
+  "settings_offline": "offline",
+  "settings_back_aria": "Back",
+  "settings_diag_issue_one": "1 issue",
+  "settings_diag_issues": "{count} issues",
+  "settings_plugins_installed_summary": "{count} installed",
+  "feat_settings_redesign_title": "Settings, redesigned",
+  "feat_settings_redesign_body": "Settings now opens as a cockpit: a section rail on the left (a full-screen list with drill-in pages on phones), aligned setting rows, and a search field — press / to jump to it — that filters sections and highlights matches. Changes still apply immediately."
 }
