+++ ui/src/lib/feature-announcements/entries/v1.45.0-settings-redesign.ts
@@ -1,0 +1,10 @@
+import type { FeatureAnnouncement } from "../../feature-announcements";
+
+const entry = {
+  id: "settings-redesign",
+  sinceVersion: "1.45.0",
+  titleKey: "feat_settings_redesign_title",
+  bodyKey: "feat_settings_redesign_body",
+} satisfies FeatureAnnouncement;
+
+export default entry;
